backend/requirements.txt
fastapi==0.115.12
pydantic==2.11.3
python-dotenv==1.1.0
# python_bcrypt==0.3.2
python_jose==3.4.0
SQLAlchemy==2.0.38
uvicorn==0.34.0
email_validator==2.2.0
psycopg2-binary==2.9.6
python-multipart==0.0.6
bcrypt==4.3.0
fastapi-utils==0.8.0
typing-inspect